+++ backend/MarketApi/MarketApi/Services/IProductService.cs
@@ -4,6 +4,7 @@ namespace MarketApi.Services
 {
     public interface IProductService
     {
+        Task<Product> CreateProductAsync(CreateProductRequest request, string walletAddress);
         Product GetProductResponse(Product product);
         ProductFile GetProductFileResponse(ProductFile productFile);
     }